pyproject.toml
[project]
name = "jobs-data"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=3.0.1",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.2",
]
